Clean code : Move add filters to Filters.jsx
